Added loop_plant to "introduce error" button
This works but display only refreshes after the loop. A timeout of 1 sec
has been added so it doesn't run too quickly (but it still only
refreshes at the end)

An extra argument was added to advance_plant(), on a button press the
"event" is also passed as an argument, this is defaulted to '0' so that
`advance_plant()` can still be used in other places in the code.